feat(Maintenance Frontend Sync): Restore and align maintenance UI with updated repository API methods
- Restored and re-integrated getMaintenanceForVehicle and getAllMaintenance repository methods for frontend consumption
- Reintroduced MaintenanceRepository callback interfaces to support asynchronous UI data loading
- Fixed MaintenanceListActivity compilation issues by aligning method calls with updated repository structure
- Fixed ReportActivity data loading by restoring getAllMaintenance callback-based API usage
- Ensured MaintenanceListActivity correctly loads, displays, and refreshes maintenance records per vehicle
- Ensured ReportActivity correctly loads full maintenance dataset for reporting view
- Maintained Retrofit-based API integration while restoring required UI-facing repository contract methods
- Verified end-to-end frontend flow for maintenance list, edit, delete, and report generation screens

+++ app/src/main/java/com/dylanclarke/FleetManagementApp/ui/MaintenanceDetailActivity.java
@@ -56,7 +56,7 @@ public class MaintenanceDetailActivity extends AppCompatActivity {
         }
 
         if (getIntent().hasExtra("maintenanceId")) {
-            maintenanceId = getIntent().getIntExtra("maintenanceId", -1);
+            maintenanceId = (int) getIntent().getLongExtra("maintenanceId", -1L);
             loadMaintenance();
         }
